// projects/ngssm-store/visibility/src/actions/ngssm-visibility-action-type.ts
export enum NgssmVisibilityActionType {
  defineElementsGroup = '[NgssmVisibilityActionType] defineElementsGroup',
  toggleElementVisibility = '[NgssmVisibilityActionType] toggleElementVisibility',
  hideElement = '[NgssmVisibilityActionType] hideElement',
  showElement = '[NgssmVisibilityActionType] showElement'
}